requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
peft>=0.10.0
ultralytics>=8.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.11.0
tqdm>=4.65.0
gradio>=4.0.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0